backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
groq==0.5.0
google-generativeai==0.7.2
pydantic==2.6.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
